tools: Move validators to validators.py
