+++ devito/tools/__init__.py
@@ -1,2 +1,3 @@
 from devito.tools.utils import *  # noqa
 from devito.tools.mixins import *  # noqa
+from devito.tools.validators import *  # noqa
